Fix range of numbers to test for primality.
Prime numbers below the range may be a factor of the value used for
testing.  This has the effect of misidentifying some prime numbers as
composite.  Composite numbers above the range may have a prime factor
greater than the largest factor of the number used for testing.  This
has the effect of misidentifying some composite numbers as prime.

Instead of allowing the user to chose the start and end of the range,
the user now specifies a number to search around.  A valid range is
computed such that the chosen number is within the range.